Add Algebra.Graph.Label
