+++ src/Algebra/Graph/Labelled.hs
@@ -16,11 +16,8 @@
 -----------------------------------------------------------------------------
 module Algebra.Graph.Labelled (
     -- * Algebraic data type for edge-labeleld graphs
-    Dioid (..), Graph (..), UnlabelledGraph, empty, vertex, edge, overlay,
-    connect, connectBy, (-<), (>-),
-
-    -- * Distances
-    Distance (..),
+    Graph (..), UnlabelledGraph, empty, vertex, edge, overlay, connect,
+    connectBy, (-<), (>-),
 
     -- * Operations
     edgeLabel
@@ -28,39 +25,9 @@ module Algebra.Graph.Labelled (
 
 import Prelude ()
 import Prelude.Compat
-import Data.Set (Set)
 
+import Algebra.Graph.Label
 import qualified Algebra.Graph.Class as C
-import qualified Data.Set as Set
-
--- | A bounded join semilattice, satisfying the following laws:
---
--- * Commutativity:         x \/ y == y \/ x
--- * Associativity:  x \/ (y \/ z) == (x \/ y) \/ z
--- * Identity:           x \/ zero == x
--- * Idempotence:           x \/ x == x
---
-class Semilattice a where
-    zero :: a
-    (\/) :: a -> a -> a
-
--- | Dioid is an idempotent semiring:
---
--- *   Associativity:  x /\ (y /\ z) == (x /\ y) /\ z
--- *   Identity:            x /\ one == x
---                          one /\ x == x
--- *   Annihilating zero:  x /\ zero == zero
---                         zero /\ x == zero
---
--- *   Distributivity: x /\ (y \/ z) == x /\ y \/ x /\ z
---                     (x \/ y) /\ z == x /\ z \/ y /\ z
---
-class Semilattice a => Dioid a where
-    one  :: a
-    (/\) :: a -> a -> a
-
-infixl 6 \/
-infixl 7 /\
 
 -- | Edge-labelled graphs, where the type variable @e@ stands for edge labels.
 data Graph e a = Empty
@@ -120,50 +87,4 @@ edgeLabel x y (Connect e g h) = edgeLabel x y g \/ edgeLabel x y h \/ new
     new | x `elem` g && y `elem` h = e
         | otherwise                = zero
 
-instance Semilattice Bool where
-    zero = False
-    (\/) = (||)
-
-instance Dioid Bool where
-    one  = True
-    (/\) = (&&)
-
 type UnlabelledGraph a = Graph Bool a
-
-data Distance a = Finite a | Infinite deriving (Eq, Ord, Show)
-
-instance (Ord a, Num a) => Num (Distance a) where
-    fromInteger = Finite . fromInteger
-
-    Infinite + _ = Infinite
-    _ + Infinite = Infinite
-    Finite x + Finite y = Finite (x + y)
-
-    Infinite * _ = Infinite
-    _ * Infinite = Infinite
-    Finite x * Finite y = Finite (x * y)
-
-    negate _ = error "Negative distances not allowed"
-
-    signum (Finite 0) = 0
-    signum _ = 1
-
-    abs = id
-
-instance Ord a => Semilattice (Distance a) where
-    zero = Infinite
-
-    Infinite \/ x = x
-    x \/ Infinite = x
-    Finite x \/ Finite y = Finite (min x y)
-
-instance (Num a, Ord a) => Dioid (Distance a) where
-    one  = Finite 0
-
-    Infinite /\ _ = Infinite
-    _ /\ Infinite = Infinite
-    Finite x /\ Finite y = Finite (x + y)
-
-instance Ord a => Semilattice (Set a) where
-    zero = Set.empty
-    (\/) = Set.union
